+++ app/src/main/java/com/fibelatti/photowidget/configure/PhotoWidgetConfigureScreen.kt
@@ -428,7 +428,8 @@ private fun PhotoWidgetConfigureContent(
                                 id = when (photoWidget.cycleMode.loopingInterval.timeUnit) {
                                     TimeUnit.SECONDS -> R.plurals.photo_widget_configure_interval_current_seconds
                                     TimeUnit.MINUTES -> R.plurals.photo_widget_configure_interval_current_minutes
-                                    else -> R.plurals.photo_widget_configure_interval_current_hours
+                                    TimeUnit.HOURS -> R.plurals.photo_widget_configure_interval_current_hours
+                                    else -> R.plurals.photo_widget_configure_interval_current_days
                                 },
                                 count = photoWidget.cycleMode.loopingInterval.repeatInterval.toInt(),
                                 photoWidget.cycleMode.loopingInterval.repeatInterval,