+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable user-facing changes to this app will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* It's now possible to set a cycle interval in days
+* Added more help articles for troubleshooting
+
 ## [v1.14.0] - 2024-09-14
 
 ### Added